feat(dashboard): AsyncState<T> infrastructure + 2 component migrations (#4189)
* feat(dashboard): add AsyncState<T> infrastructure and migrate 2 components

Introduce discriminated union AsyncState<T> and createAsyncResource<T>()
to replace the scattered 3-signal pattern (data/loading/error + manual
request dedup) found in 22+ dashboard components.

New files:
- lib/async-state.ts: type + resource factory + type guards (16 tests)
- components/common/async-container.ts: render wrapper for AsyncState

Migrated:
- feature-health.ts: full migration with AsyncContainer
- harness-health.ts: resource migration with direct state read
  (AsyncContainer skipped due to vi.resetModules signal isolation)


* refactor(dashboard): migrate 7 components to AsyncResource (batch 1)

activity-swimlane, activity-graph, tool-metrics, collaboration-evidence,
logs, server-config, agent-profile — all converted from 3-signal pattern
to createAsyncResource<T>(). Net -19 lines, eliminates manual request
dedup in each file.


* refactor(dashboard): migrate 5 state modules to AsyncResource (batch 2)

keeper-spawn-state, tool-state, governance-signals/actions,
flow-control-state, keeper-config-panel — converted from 3-signal
pattern to createAsyncResource<T>(). Backward-compatible computed
signals exposed where consumers depend on them.

14 files correctly skipped (UI selection/form signals, not async trios).


* refactor(dashboard): split autoresearch.ts god file into 3 modules

879 lines → 3 focused modules:
- autoresearch-state.ts (156): AsyncResource for loops, manual signals
  for detail (race-condition guard), actions, refresh
- autoresearch-form.ts (261): 13 form signals, submit handler,
  StartAutoresearchForm dialog
- autoresearch.ts (510): UI components + main Autoresearch component

Dependency: autoresearch → state, form → state. No cycles.
All exports preserved for external consumers.


* refactor(dashboard): split memory.ts god file into 3 modules

876 lines → 3 focused modules:
- memory-state.ts (317): signals, data ops, vote/delete/comment handlers
- memory-post-detail.ts (262): PostDetail, CommentThread, CommentForm
- memory.ts (404): Memory main, PostCard, SortBar, NewPostForm

Dependency: memory → state + post-detail, post-detail → state. No cycles.


* fix(dashboard): handle sync throws in AsyncResource.load and use theme variables

Wrap fn() call in try-catch so synchronous throws transition to
error state instead of leaving state stuck in loading with null
inflight. Replace hardcoded Tailwind red colors in AsyncContainer
error UI with --bad CSS variables for theme consistency.


* refactor(dashboard): split harness-health.ts god file into 3 modules

853 lines → 3 focused modules:
- harness-health-state.ts (269): types, AsyncResource, SSE handler,
  timer management, data loading
- harness-health-sections.ts (369): helper functions, StatusPill,
  StatCard, GateChart, section components
- harness-health.ts (247): HarnessHealth main component only

Dependency: harness-health → state + sections, sections → state types.


* fix(dashboard): address review findings from /simplify

- async-state: add generation counter to prevent stale writes after
  reset() — reset increments gen, in-flight .then/.catch skips if
  gen has changed
- async-container: use LoadingState (animate-pulse) instead of
  EmptyState for loading/idle states
- activity-graph/swimlane: remove unnecessary reset() before load()
  that defeated request deduplication
- harness-health-sections: rename statusLabel → railStatusLabel to
  avoid name collision with lib/status-label.ts


* fix(dashboard): address 2nd-pass review findings

- governance-actions: read resource state directly after load() instead
  of relying on catch block (load() swallows errors internally)
- agent-profile: only reset() when agent name changes, preventing
  UI loading flash on same-agent refresh
- logs: replace useSignalEffect with useEffect to prevent re-entrance
  loop from reset() writing to tracked signal
- Extract StatCard to common/stat-card.ts, removing duplication between
  feature-health.ts and harness-health-sections.ts


* fix(dashboard): move governance post-load into callback + add reset tests

- governance-actions: move selection/detail logic inside load() callback
  so concurrent callers are deduplicated (side effects run once)
- async-state.test: add 2 tests for reset-while-inflight behavior —
  verifies generation counter discards stale writes after reset()


* fix(dashboard): restore loading indicator during room strategy mutation

setRoomStrategy mutation phase was missing loading state after
AsyncResource migration. Add roomStrategyMutating signal so
roomStrategyLoading computed covers both fetch and mutation.


* fix(dashboard): move PostDetail data fetch from render to useEffect

Render-time signal mutation (loadPostDetail inside render body) is
unsafe and can cause unnecessary re-renders. Move to useEffect with
post.id dependency.


---------
