+++ dashboard/src/lib/async-state.ts
@@ -1,0 +1,104 @@
+// Async state management — eliminates the 3-signal pattern (data/loading/error)
+// scattered across 22+ dashboard components.
+//
+// Before: const data = signal<T | null>(null)
+//         const loading = signal(false)
+//         const error = signal<string | null>(null)
+//         let request: Promise<void> | null = null
+//
+// After:  const resource = createAsyncResource<T>()
+
+import { signal, type Signal } from '@preact/signals'
+
+// ── Discriminated union ──
+
+type Idle = { readonly status: 'idle' }
+type Loading = { readonly status: 'loading' }
+type Loaded<T> = { readonly status: 'loaded'; readonly data: T }
+type Failed = { readonly status: 'error'; readonly message: string }
+
+export type AsyncState<T> = Idle | Loading | Loaded<T> | Failed
+
+// ── Constructors ──
+
+export const idle: Idle = { status: 'idle' }
+export const loading: Loading = { status: 'loading' }
+export function loaded<T>(data: T): Loaded<T> { return { status: 'loaded', data } }
+export function failed(message: string): Failed { return { status: 'error', message } }
+
+// ── Type guards ──
+
+export function isLoaded<T>(state: AsyncState<T>): state is Loaded<T> {
+  return state.status === 'loaded'
+}
+
+export function isLoading<T>(state: AsyncState<T>): state is Loading {
+  return state.status === 'loading'
+}
+
+export function isFailed<T>(state: AsyncState<T>): state is Failed {
+  return state.status === 'error'
+}
+
+// ── Data extraction (returns undefined for non-loaded states) ──
+
+export function getData<T>(state: AsyncState<T>): T | undefined {
+  return state.status === 'loaded' ? state.data : undefined
+}
+
+// ── Managed async resource ──
+//
+// Bundles a signal holding AsyncState<T> with request deduplication.
+// Replaces the manual `let request: Promise<void> | null` pattern.
+
+export interface AsyncResource<T> {
+  readonly state: Signal<AsyncState<T>>
+  load(fn: () => Promise<T>): Promise<void>
+  reset(): void
+}
+
+export function createAsyncResource<T>(): AsyncResource<T> {
+  const state = signal<AsyncState<T>>(idle)
+  let inflight: Promise<void> | null = null
+  let generation = 0
+
+  return {
+    state,
+
+    load(fn: () => Promise<T>): Promise<void> {
+      if (inflight) return inflight
+
+      const gen = ++generation
+      state.value = loading
+
+      let promise: Promise<T>
+      try {
+        promise = fn()
+      } catch (e) {
+        if (gen === generation) {
+          state.value = failed(e instanceof Error ? e.message : String(e))
+        }
+        return Promise.resolve()
+      }
+
+      inflight = promise
+        .then(data => {
+          if (gen === generation) state.value = loaded(data)
+        })
+        .catch(e => {
+          if (gen === generation) state.value = failed(e instanceof Error ? e.message : String(e))
+        })
+        .finally(() => {
+          if (gen === generation) inflight = null
+        })
+
+      return inflight
+    },
+
+    reset(): void {
+      ++generation
+      state.value = idle
+      inflight = null
+    },
+  }
+}
